requirements.txt
fastapi
uvicorn[standard]
pydantic
PyYAML
requests
chromadb
sentence-transformers
torch
langchain
langchain-community
unstructured
python-docx
pandas
openpyxl
groq
langgraph
python-multipart